extract hermione reporter utils
